Switched

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,14 +4,14 @@ import java.math.BigDecimal;
 
 public class Account {
     private int account_id;
-    private BigDecimal balance;
+    private Double balance;
     private int user_id;
 
     public Account() {
     }
 
 
-    public Account(int account_id, BigDecimal balance, int user_id) {
+    public Account(int account_id, Double balance, int user_id) {
         this.account_id = account_id;
         this.balance = balance;
         this.user_id = user_id;
@@ -25,11 +25,11 @@ public class Account {
         this.account_id = account_id;
     }
 
-    public BigDecimal getBalance() {
+    public Double getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
